Agents/RAG Service/requirements.txt
# MCP Protocol Dependencies
httpx>=0.24.0
uvicorn>=0.23.0
starlette>=0.27.0
fastapi>=0.104.0
click>=8.1.0
pydantic>=2.0.0

# LangChain and LangGraph
langchain>=0.3.15
langchain-core>=1.0.7
langchain-google-genai>=3.1.0
langchain-community>=0.3.0  # For document loaders

# Supabase and PostgreSQL
supabase>=2.24.0
psycopg[binary]>=3.1.0  # For direct pgvector operations
psycopg-pool>=3.1.0

# Document Processing
pypdf>=3.17.0  # PDF processing
python-dotenv>=1.2.1
beautifulsoup4>=4.12.0  # HTML parsing for hybrid chunking

# Machine Learning & NLP
numpy>=1.24.0  # For vector operations
scikit-learn>=1.3.0  # Optimized cosine similarity for semantic chunking

